% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/now-package.R
\docType{package}
\name{now-package}
\alias{now}
\alias{now-package}
\title{now: Remove Exported Functions Depending On Lifecycle}
\description{
Remove "experimental", "superseded", "deprecated", "defunct", "questioning", or "reexports" from the exported functions of a package, so these don't get in the way.
}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@cynkra.com}

}
\keyword{internal}
